// src/data/type/country-types.ts
export type CountryTypes = {
  altSpellings?: Array<String>;
  area?: Number;
  borders?: Array<String>;
  capital?: Array<String>;
  capitalInfo?: any;
  car?: any;
  cca2?: String;
  cca3?: String;
  ccn3?: String;
  cioc?: String;
  coatOfArms?: any;
  continents?: Array<String>;
  currencies?: any;
  demonyms?: any;
  fifa?: String;
  flag?: String;
  flags?: any;
  gini?: any;
  idd?: any;
  independent?: boolean;
  landlocked?: boolean;
  languages?: any;
  latlng?: Array<Number>;
  maps?: any;
  name?: NameTypes;
  population?: Number;
  postalCode?: any;
  region?: String;
  startOfWeek?: String;
  status?: String;
  subregion?: String;
  timezones?: Array<String>;
  tld?: Array<String>;
  translations?: any;
  unMember?: boolean;
};

type NameTypes = {
  common?: String;
  nativeName?: any;
  official?: String;
};